Move Repositories to new namespace, change class names
Repository class names now include Json to diffrentiante them from future impelmentations.
