+++ TutorLizard.BusinessLogic/Data/Repositories/Json/JsonRepositoryBase.cs
@@ -1,15 +1,8 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Text.Json;
-using TutorLizard.BusinessLogic.Models;
-using System.Diagnostics;
+﻿using System.Text.Json;
 
-namespace TutorLizard.BusinessLogic.Data
+namespace TutorLizard.BusinessLogic.Data.Repositories.Json
 {
-    public abstract class RepositoryBase<T>
+    public abstract class JsonRepositoryBase<T>
     {
         protected List<T> _data = new();
         protected abstract string FilePath { get; }
@@ -58,6 +51,6 @@ namespace TutorLizard.BusinessLogic.Data
                 Directory.CreateDirectory(directoryPath);
 
             File.WriteAllText(fullPath, jsonData);
-        }     
+        }
     }
 }
