+++ TutorLizard.BusinessLogic/TutorLizard.BusinessLogic.csproj
@@ -6,7 +6,4 @@
     <Nullable>enable</Nullable>
   </PropertyGroup>
 
-  <ItemGroup>
-  </ItemGroup>
-
 </Project>
